Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -3,7 +3,7 @@
 
 
 int main() {
-    int i;
+    float i;
     printf();
     scanf("%f\n" &i);
      return 0;
